tests/util: implement Piglit signal handler.
The signal handler is extremely useful for distributed testing
where the best course of action is to ask the end user to
simply submit the piglit main results file. This means that the
generated reports will now include call stacks.

Note : I have not tested windows functionality yet, but the basic
premis is in place, and it appears that windows also has support
for signal.h

Note 2: On linux the SIGABRT is not always handled as expected.
This specific signal requires an alternative ability to generate
call stacks.