Minor Change for geometry shaders.
For GLSL 1.20 to 3.30: After reading the spec, it appears that the
following declaration for the length function of an array is invalid.

gl_PositionIn.length

The correct syntax is to instead make calls using the syntax...
gl_PositionIn.length()

see Section 4.1.9 ( arrays ) of the glsl spec.

NOTE: The default behavior for nvidia drivers is to allow for cg
language extensions to be avaialble by default. Please keep this
in mind because the "gl_PositionIn.length" declaration is valid
when using CG language constructions.

For GLSL 1.10: There is no length operator. However, this is no
problem because gl_VerticesIn is defined as a constant integer
in the ARB_geometry_shader4 spec.