modified GLX_SGI_video_sync to use XRandr to query current refresh rate.
NOTE: The execution test will run a variable length of time dependent
on the detected refreash rate. If the refresh rate is determined, the
result will be that the execution test will run 5 sub tests with an
expected vblank count for each test being 5 times the refresh rate.
For example, If thee current refresh rate is 60hz, the each subtest will
be to 300 vblanks.  If the refresh rate is not determined, the subtest
will run until a total of 60 vblanks have passed. If there is problems
with the vblank system and how the program is ran, the test will detect
that vblanks are missed and thus fail.