cmake: ignore the "CMakeFiles" folder in generated_tests installer
The CMakeFiles folder contains nothing that is actually used for the
runtime of tests. So it is safe to completely ignore this path when
actually installing or packaging piglit. The contents of CMakeFiles
is nothing but internal CMake junk. This is also ignored when creating
git commits.

V2: modified commit log to be a little bit clearer.