piglit-util-gl-common: Add Function to clear extensions list.
The default handling of context fallback does not follow the
pre-defined behavior that the OpenGL Specification requires.
The OpenGL Specifications specifically state that queries to
glGetString ( and glGetStringi ) is to be the current GL
connection. When you change OpenGL Contexts, the GL connection
also changes, so this means that all previous strings should be
considered as invalid.

Chapter 6 of the OpenGL 1.1 Specification on page 147 states:
   returns a pointer [...] of the current GL Connection.

Section 6.1.11 of OpenGL 2.0 and 2.1 Specifications states:
   returns a pointer [...] of the current GL connection

Section 6.1.6 of OpenGL 3.3 Core Profile Specification states:
   String queries return pointers [...] of the current GL context.

Section 6.1.5 of OpenGL ES 1.1 Specification States:
   returns a pointer [...] of the current GL connection.

Using this wording, one can view the Current Context as the
GL Connection and when you change OpenGL Connections, the
result is that the old value is now invalidated.

V2: applied the changes suggested by Chad to improve this function.
    In my previous submission of this I forgot to make note of this.
    This commit changed the new function to:

    piglit_gl_reinitialize_extensions

V3:Changed log of this comit to reference exact OpenGL ES and OpenGL
    Specificatoin sections.