util: Implement signal handler support.
* Non-Posix compliant platforms - Only a stub for a signal handler exists.

* Posix platforms - Add support for simple backtrace support when a crash is
detected. This support does not interfere with valgrind and gdb testing.

When running a program in the debugger ( for example, GDB ), the singal handler
is only handled after you resume the program using the the signal handler.

v2: changed all of the signal handler printing functions to use stderr.
The only report that goes to stdout is the status report for piglit.

v3: Made a few more changes.
* Posix: Increased stack trace to 64 frames.
* Posix: Added a few more signals to catch.