util: Implement signal handler on windows.
This signal handler is built using dbghelp support. This is found in most
versions of windows, and should work great on Windows XP or newer. I
have tested this against Windows 7 using MSVC and there is no major
problems. Also, as a benefit, this will automatically generate the
backtrace information within the piglit results file.

v2: changed all of the signal handler printing functions to use stderr.
The only report that goes to stdout is the status report for piglit.